Calculating Profits
Calculate the total sale of material to one company. Selling to this single company all the material that can be sold to them until material is gone, or almost gone (without going over the amount of material we have of course). That is assuming that we can sell to a single company more than once.