Take in data from a separate file
This program was made assuming that the table/list of companies to sell to is not information that exists within the program initially. The program reads the data from a separate file (that the user can direct it to) and puts it into a list which then goes through the program, producing a ranked list of the buying companies from most profitable to least profitable.