Printing final solution
Assuming you can only sell to a company once, the program first calculates the prices that each company (info gathered from a separate file) will pay per unit and organizes a list of the companies, from the company paying the most per unit to the company paying the least per unit. The program then takes into account the amount of material the selling company has to offer (user input) and calculates how many of the top contenders the selling company will be able to profit from. Finally, the program finds the total amount of profit that can be gained based on how much the selling company purchased the material for (user input) and how much the selling company can gather from the highest paying buyer companies and compiles the information into two lists.